+++ package/Editor/ValidationSuite/UI/PackageValidationWindow.cs
@@ -7,38 +7,8 @@ namespace UnityEditor.PackageManager.ValidationSuite.UI
 {
     internal class PackageValidationWindow
     {
-        private static string resultsFilePath = "ValidationSuiteResults.txt";
         private static string ResultsDisplay = "";
         private static bool ShowResultsDlg = false;
-
-        public static bool RunValidationTestsSuite()
-        {
-            ResultsDisplay = "";
-            var packagePath = FindFirstPackagePath();
-
-            if (string.IsNullOrEmpty(packagePath))
-            {
-                return false;
-            }
-
-            try
-            {
-                // Clear output file content
-                File.WriteAllText(resultsFilePath, string.Format("Validation Suite Results for package \"{0}\"\r\n\r\n", packagePath));
-
-                var package = new VettingContext();
-                package.Initialize(packagePath);
-
-                ValidationSuite testSuite = new ValidationSuite(SingleTestCompletedDelegate, AllTestsCompletedDelegate, package);
-
-                testSuite.RunAsync();
-            }
-            catch(Exception e)
-            {
-                File.AppendAllText(resultsFilePath, string.Format("\r\nTest Setup Error: \"{0}\"\r\n", e.Message));
-            }
-            return true;
-        }
 
         [MenuItem("internal:Project/Packages/Validate")]
         internal static void ShowPackageManagerWindow()
@@ -60,7 +30,7 @@ namespace UnityEditor.PackageManager.ValidationSuite.UI
             try
             {
                 // Clear output file content
-                File.WriteAllText(resultsFilePath, string.Format("Validation Suite Results for package \"{0}\"\r\n\r\n", packagePath));
+                File.WriteAllText(ValidationSuite.resultsFilePath, string.Format("Validation Suite Results for package \"{0}\"\r\n\r\n", packagePath));
 
                 var package = new VettingContext();
                 package.Initialize(packagePath);
@@ -71,7 +41,7 @@ namespace UnityEditor.PackageManager.ValidationSuite.UI
             }
             catch(Exception e)
             {
-                File.AppendAllText(resultsFilePath, string.Format("\r\nTest Setup Error: \"{0}\"\r\n", e.Message));
+                File.AppendAllText(ValidationSuite.resultsFilePath, string.Format("\r\nTest Setup Error: \"{0}\"\r\n", e.Message));
             }
         }
 
@@ -88,17 +58,17 @@ namespace UnityEditor.PackageManager.ValidationSuite.UI
         // This is called after every tests completes
         private static void SingleTestCompletedDelegate(IValidationTestResult testResult)
         {
-            File.AppendAllText(resultsFilePath, string.Format("\r\nTest: \"{0}\"\r\nResult: {1}\r\n", testResult.ValidationTest.TestName, testResult.TestState));
+            File.AppendAllText(ValidationSuite.resultsFilePath, string.Format("\r\nTest: \"{0}\"\r\nResult: {1}\r\n", testResult.ValidationTest.TestName, testResult.TestState));
             ResultsDisplay += string.Format("{0} - {1}\n", testResult.TestState, testResult.ValidationTest.TestName);
             if (testResult.TestOutput.Any())
             {
-                File.AppendAllText(resultsFilePath, string.Join("\r\n", testResult.TestOutput.ToArray()) + "\r\n");
+                File.AppendAllText(ValidationSuite.resultsFilePath, string.Join("\r\n", testResult.TestOutput.ToArray()) + "\r\n");
             }
         }
 
         private static void AllTestsCompletedDelegate(TestState testRunState)
         {
-            File.AppendAllText(resultsFilePath, "\r\nAll Done!  Result = " + testRunState);
+            File.AppendAllText(ValidationSuite.resultsFilePath, "\r\nAll Done!  Result = " + testRunState);
             ShowResultsDlg = true;
         }