+++ package/Editor/ValidationSuite/ValidationTests/AssemblyDefinitionValidation.cs
@@ -6,6 +6,7 @@ using UnityEditor.PackageManager.ValidationSuite;
 
 namespace UnityEditor.PackageManager.ValidationSuite.ValidationTests
 {
+    /** Skip it for now
     internal class AssemblyDefinitionValidation : BaseValidation
     {
         internal class AssemblyDefinitionData {
@@ -143,4 +144,5 @@ namespace UnityEditor.PackageManager.ValidationSuite.ValidationTests
             }
         }
     }
+    **/
 }