+++ package/CHANGELOG.md
@@ -4,5 +4,10 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.1.1] - 2018-02-06
+### Changes
+- Add static public method to run test from script
+- Add synchronous call for Cloud script
+
 ## [0.1.0] - 2017-12-20
 ### This is the first release of *Unity Package Validation Suite*.